requirements.txt
flask
flask-cors
mysql-connector-python
web3
gunicorn